# Backend/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Volcanoes Around the World! v2.0",
        "description": "This Swagger API document exposes a number of REST endpoints which implement CRUD operations on a database containing publicly available data about volcanoes collated by the Smithsonian Institution's Global Volcanism Program. The API endpoints and their usage are described in detail below.",
        "contact": {
            "email": "n11078472@qut.edu.au"
        }
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token only:"
            }
        }        
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/review/{id}": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Post a review of a volcano.",
                "description": "Post a rating and a comment about a volcano, that other's will be able to view. A rating is required, a comment is optional.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "description": "The ID of the volcano to be reviewed"
                    }
                ],
                "requestBody": {
                    "description": "An object containing the rating out of 5 (1, 2, 3, 4, or 5), and optionally a comment.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Had a great time visiting!"
                                    }
                                },
                                "required": [
                                    "rating"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review successfully posted and added to the system. An object with a success message is returned.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review successfully added"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid input data or missing fields. Click the examples dropdown to see the possible error responses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "MissingRating": {
                                        "summary": "MissingRating",
                                        "value": {
                                            "error": true,
                                            "message": "Request body incomplete: a rating is required"
                                        }
                                    },
                                    "InvalidRatingType": {
                                        "summary": "InvalidRatingType",
                                        "value": {
                                            "error": true,
                                            "message": "Invalid rating: must be 1, 2, 3, 4, or 5"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized errors. Click on examples dropdown to see possible error responses.",
                        "content": {
                            "application/json": {
                                "type": "object",
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "MissingAuthHeader": {
                                        "summary": "MissingAuthHeader",
                                        "value": {
                                            "error": true,
                                            "message": "Authorization header not found. Must be logged in to post a review"
                                        }
                                    },
                                    "TokenExpired": {
                                        "summary": "TokenExpired",
                                        "value": {
                                            "error": true,
                                            "message": "JWT token has expired"
                                        }
                                    },
                                    "InvalidJWT": {
                                        "summary": "InvalidJWT",
                                        "value": {
                                            "error": true,
                                            "message": "Invalid JWT token"
                                        }
                                    },
                                    "MalformedAuthHeader": {
                                        "summary": "MalformedAuthHeader",
                                        "value": {
                                            "error": true,
                                            "message": "Authorization header is malformed"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - User does not exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server / database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get all reviews of a specific volcano.",
                "description": "Returns a list of review objects for the specified volcano. Each review includes a star rating out of 5, a comment, and the first and last name of the user that posted the review.",
                "security": [],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "description": "The ID of the volcano to be reviewed"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of review objects associated to the speicifed volcano. Each object contains a rating (1 to 5), a comment, and the firstName and lastName of the user who posted the review. An example is shown below.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "rating": {
                                                "type": "integer",
                                                "description": "Rating of the review (1 to 5)"
                                            },
                                            "comment": {
                                                "type": [
                                                    "string",
                                                    "null"
                                                ],
                                                "description": "Comment provided by the user (can be null)"
                                            },
                                            "firstName": {
                                                "type": "string",
                                                "description": "First name of the user who posted the review"
                                            },
                                            "lastName": {
                                                "type": "string",
                                                "description": "Last name of the user who posted the review"
                                            }
                                        },
                                        "example": [
                                            {
                                                "rating": 1,
                                                "comment": "It erupted while I was taking a selfie. Only just made it out alive.",
                                                "firstName": "Sam",
                                                "lastName": "van Zuylen"
                                            },
                                            {
                                                "rating": 5,
                                                "comment": "Loved it!",
                                                "firstName": "Michael",
                                                "lastName": "Jordan"
                                            },
                                            {
                                                "rating": 4,
                                                "comment": "Was pretty great",
                                                "firstName": "Timothy",
                                                "lastName": "Chappell"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - no reviews found for the specified volcano.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No reviews found for this volcano"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server / database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}